# docs/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/api/v1/households/{householdId}/recipes/{recipeId}" : {
      "get" : {
        "tags" : [ "recipes-controller" ],
        "operationId" : "get",
        "parameters" : [ {
          "name" : "householdId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "recipeId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RecipeDto"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "recipes-controller" ],
        "operationId" : "update",
        "parameters" : [ {
          "name" : "householdId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "recipeId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateRecipe"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RecipeDto"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "recipes-controller" ],
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "householdId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "recipeId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
...
    "/api/v1/households/{householdId}/recipes" : {
      "get" : {
        "tags" : [ "recipes-controller" ],
        "operationId" : "list_1",
        "parameters" : [ {
          "name" : "householdId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/RecipeDto"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "recipes-controller" ],
        "operationId" : "create_1",
        "parameters" : [ {
          "name" : "householdId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateRecipe"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RecipeDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/households/{householdId}/recipes/search" : {
      "post" : {
        "tags" : [ "recipes-controller" ],
        "operationId" : "search",
        "parameters" : [ {
          "name" : "householdId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "tag",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "ingredient",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 24
          }
        }, {
          "name" : "sort",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "default" : "name"
          }
        }, {
          "name" : "direction",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "default" : "asc"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/FilterRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RecipePageDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "CreateRecipe" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string",
            "minLength" : 1
          },
          "servings" : {
            "type" : "integer",
            "format" : "int32"
          },
          "notes" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "steps" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "tags" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "ingredients" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/IngredientDto"
            }
          },
          "imageMimeType" : {
            "type" : "string"
          },
          "imageData" : {
            "type" : "string"
          }
        },
        "required" : [ "name" ]
      },
      "IngredientDto" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          },
          "quantity" : {
            "type" : "number",
            "format" : "double"
          },
          "unit" : {
            "type" : "string"
          },
          "category" : {
            "type" : "string"
          }
        }
      },
      "RecipeDto" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "householdId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          },
          "servings" : {
            "type" : "integer",
            "format" : "int32"
          },
          "notes" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "steps" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "tags" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "imageMimeType" : {
            "type" : "string"
          },
          "imageData" : {
            "type" : "string"
          },
          "ingredients" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/IngredientDto"
            }
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "FilterCondition" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/FilterNode"
        }, {
          "type" : "object",
          "properties" : {
            "field" : {
              "type" : "string"
            },
            "operator" : {
              "type" : "string",
              "enum" : [ "EQ", "NE", "GT", "GTE", "LT", "LTE", "LIKE", "IN", "BETWEEN", "IS_NULL", "IS_NOT_NULL", "JSON_CONTAINS", "JSON_CONTAINED_BY", "JSON_EXISTS", "JSON_EXISTS_ANY", "JSON_EXISTS_ALL", "JSON_PATH_EQ", "JSON_PATH_LIKE", "JSON_ARRAY_CONTAINS", "JSON_ARRAY_CONTAINS_ALL", "JSON_ARRAY_CONTAINS_ANY" ]
            },
            "value" : { },
            "values" : {
              "type" : "array",
              "items" : { }
            }
          }
        } ]
      },
      "FilterGroup" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/FilterNode"
        }, {
          "type" : "object",
          "properties" : {
            "combinator" : {
              "type" : "string",
              "enum" : [ "AND", "OR" ],
              "writeOnly" : true
            },
            "filters" : {
              "type" : "array",
              "items" : {
                "oneOf" : [ {
                  "$ref" : "#/components/schemas/FilterCondition"
                }, {
                  "$ref" : "#/components/schemas/FilterGroup"
                } ]
              },
              "writeOnly" : true
            },
            "and" : {
              "type" : "array",
              "items" : {
                "oneOf" : [ {
                  "$ref" : "#/components/schemas/FilterCondition"
                }, {
                  "$ref" : "#/components/schemas/FilterGroup"
                } ]
              }
            },
            "or" : {
              "type" : "array",
              "items" : {
                "oneOf" : [ {
                  "$ref" : "#/components/schemas/FilterCondition"
                }, {
                  "$ref" : "#/components/schemas/FilterGroup"
                } ]
              }
            }
          }
        } ]
      },
      "FilterNode" : { },
      "FilterRequest" : {
        "type" : "object",
        "properties" : {
          "filter" : {
            "oneOf" : [ {
              "$ref" : "#/components/schemas/FilterCondition"
            }, {
              "$ref" : "#/components/schemas/FilterGroup"
            } ]
          }
        },
        "required" : [ "filter" ]
      },
      "RecipePageDto" : {
        "type" : "object",
        "properties" : {
          "content" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/RecipeDto"
            }
          },
          "totalHits" : {
            "type" : "integer",
            "format" : "int64"
          },
          "hasMore" : {
            "type" : "boolean"
          }
        }
      },
...
    }
  }
}